Group F-19 split modules into skyulf/pipeline/ package
Move the orchestrator, reproducibility seal, and diagram builder into a
pipeline/ subpackage to keep the top-level package uncluttered.
pipeline/__init__.py re-exports only SkyulfPipeline, so the public
import path and pickle compatibility are preserved.
